+++ src/Elders.Pandora/RELEASE_NOTES.md
@@ -1,3 +1,7 @@
+#### 0.8.0 - 01.06.2017
+* Pandora raw keys will always have a machine name. If it is a cluster key the machine name is '*'
+* Adds a parser for a DeployedSetting
+
 #### 0.7.2 - 23.02.2017
 * Fix cluster comparison